Add files via upload
This project is about a product management of a marketplace with the data structs made in B-Tree. The code has 2 additional files that present the products on the B-Tree and the other has a complete description of witch produtc. The project has a complete B-Tree with all functions and structs.